Fix auto-start configuration for ST3215 and nooploop UWB drivers
- Add UWB1-4 serial port options to Kconfig to enable serial config generation
- Remove duplicate ST3215_CONFIG parameter definition from st3215_servo module.yaml
- Remove duplicate UWB_CONFIG parameter definition from nooploop_linktrack module.yaml
- Fix compilation error in nooploop_linktrack.cpp format string for int32_t

This resolves auto-start issues where serial drivers weren't being configured
properly due to missing Kconfig entries and duplicate parameter definitions.

Tested on:
- hkust_nxt-dual-wl-rear_default: ST3215 driver auto-starts on SRV1
- holybro_pixhawk-6x-pro-uwb_default: nooploop UWB driver auto-starts on UWB1

+++ src/drivers/uwb/nooploop_linktrack/nooploop_linktrack.cpp
@@ -144,7 +144,7 @@ void NoopLoopLinkTrack::Run()
             case 115200: speed = B115200; break;
             case 921600: speed = B921600; break;
             default:
-                PX4_WARN("Unsupported baud rate %d, using 921600", baud_param);
+                PX4_WARN("Unsupported baud rate %ld, using 921600", (long)baud_param);
                 speed = B921600;
                 break;
         }
